# Refactor unstack