--- a/REQUIRE
+++ b/REQUIRE
@@ -1,6 +1,6 @@
 julia 0.6
 Missings 0.2.3
-CategoricalArrays 0.3.0
+CategoricalArrays 0.3.2
 StatsBase 0.11.0
 SortingAlgorithms
 Reexport