--- a/test/dataframe.jl
+++ b/test/dataframe.jl
@@ -299,12 +299,15 @@ module TestDataFrame
 
     #Check the output of unstack
     df = DataFrame(Fish = CategoricalArray{Union{String, Missing}}(["Bob", "Bob", "Batman", "Batman"]),
-                   Key = Union{String, Missing}["Mass", "Color", "Mass", "Color"],
+                   Key = CategoricalArray{Union{String, Missing}}(["Mass", "Color", "Mass", "Color"]),
                    Value = Union{String, Missing}["12 g", "Red", "18 g", "Grey"])
     # Check that reordering levels does not confuse unstack
     levels!(df[1], ["XXX", "Bob", "Batman"])
+    levels!(df[2], ["YYY", "Color", "Mass"])
     #Unstack specifying a row column
     df2 = unstack(df, :Fish, :Key, :Value)
+    @test levels(df[1]) == ["XXX", "Bob", "Batman"] # make sure we did not mess df[1] levels
+    @test levels(df[2]) == ["YYY", "Color", "Mass"] # make sure we did not mess df[2] levels
     #Unstack without specifying a row column
     df3 = unstack(df, :Key, :Value)
     #The expected output, XXX level should be dropped as it has no rows with this key
@@ -321,6 +324,52 @@ module TestDataFrame
     #This changes the expected result
     df4[1,:Mass] = missing
     @test df2 ≅ df4
+
+    #The same as above but without CategoricalArray
+    df = DataFrame(Fish = ["Bob", "Bob", "Batman", "Batman"],
+                   Key = ["Mass", "Color", "Mass", "Color"],
+                   Value = ["12 g", "Red", "18 g", "Grey"])
+    #Unstack specifying a row column
+    df2 = unstack(df, :Fish, :Key, :Value)
+    #Unstack without specifying a row column
+    df3 = unstack(df, :Key, :Value)
+    #The expected output, XXX level should be dropped as it has no rows with this key
+    df4 = DataFrame(Fish = ["Batman", "Bob"],
+                    Color = ["Grey", "Red"],
+                    Mass = ["18 g", "12 g"])
+    @test df2 ≅ df4
+    @test typeof(df2[:Fish]) <: Vector{String}
+    # first column stays as CategoricalArray in df3
+    @test df3 == df4
+    #Make sure unstack works with missing values at the start of the value column
+    allowmissing!(df, :Value)
+    df[1,:Value] = missing
+    df2 = unstack(df, :Fish, :Key, :Value)
+    #This changes the expected result
+    allowmissing!(df4, :Mass)
+    df4[2,:Mass] = missing
+    @test df2 ≅ df4
+
+    # test empty set of grouping variables
+    @test_throws ArgumentError unstack(df, Int[], :Key, :Value)
+    @test_throws ArgumentError unstack(df, Symbol[], :Key, :Value)
+    @test_throws KeyError unstack(stack(DataFrame(rand(10, 10))))
+
+    # test missing value in grouping variable
+    mdf = DataFrame(id=[missing,1,2,3], a=1:4, b=1:4)
+    @test unstack(melt(mdf, :id))[1:3,:] == sort(mdf)[1:3,:]
+    @test unstack(melt(mdf, :id))[:,2:3] == sort(mdf)[:,2:3]
+
+    # test more than one grouping column
+    wide = DataFrame(id = 1:12,
+                     a  = repeat([1:3;], inner = [4]),
+                     b  = repeat([1:4;], inner = [3]),
+                     c  = randn(12),
+                     d  = randn(12))
+
+    long = stack(wide)
+    wide3 = unstack(long, [:id, :a], :variable, :value)
+    @test wide3 == wide[:, [1, 2, 4, 5]]
 
     df = DataFrame(A = 1:10, B = 'A':'J')
     @test !(df[:,:] === df)
@@ -350,7 +399,8 @@ module TestDataFrame
         @test udf == DataFrame(Any[Union{Int, Missing}[1, 2], Union{Int, Missing}[1, 5],
                                    Union{Int, Missing}[2, 6], Union{Int, Missing}[3, 7],
                                    Union{Int, Missing}[4, 8]], [:id, :a, :b, :c, :d])
-        @test all(isa.(udf.columns, Vector{Union{Int, Missing}}))
+        @test isa(udf[1], Vector{Int})
+        @test all(isa.(udf.columns[2:end], Vector{Union{Int, Missing}}))
         df = DataFrame(Any[categorical(repeat(1:2, inner=4)),
                            categorical(repeat('a':'d', outer=2)), categorical(1:8)],
                        [:id, :variable, :value])
@@ -359,18 +409,22 @@ module TestDataFrame
         @test udf == DataFrame(Any[Union{Int, Missing}[1, 2], Union{Int, Missing}[1, 5],
                                    Union{Int, Missing}[2, 6], Union{Int, Missing}[3, 7],
                                    Union{Int, Missing}[4, 8]], [:id, :a, :b, :c, :d])
-        @test all(isa.(udf.columns, CategoricalVector{Union{Int, Missing}}))
+        @test isa(udf[1], CategoricalVector{Int})
+        @test all(isa.(udf.columns[2:end], CategoricalVector{Union{Int, Missing}}))
     end
 
     @testset "duplicate entries in unstack warnings" begin
-        df = DataFrame(id=Union{Int, Missing}[1, 2, 1, 2], variable=["a", "b", "a", "b"], value=[3, 4, 5, 6])
+        df = DataFrame(id=Union{Int, Missing}[1, 2, 1, 2],
+                       id2=Union{Int, Missing}[1, 2, 1, 2], 
+                       variable=["a", "b", "a", "b"], value=[3, 4, 5, 6])
         @static if VERSION >= v"0.6.0-dev.1980"
-            @test_warn "Duplicate entries in unstack." unstack(df, :id, :variable, :value)
-            @test_warn "Duplicate entries in unstack at row 3." unstack(df, :variable, :value)
+            @test_warn "Duplicate entries in unstack at row 3 for key 1 and variable a." unstack(df, :id, :variable, :value)
+            @test_warn "Duplicate entries in unstack at row 3 for key (1, 1) and variable a." unstack(df, :variable, :value)
         end
         a = unstack(df, :id, :variable, :value)
+        @test a ≅ DataFrame(id = [1, 2], a = [5, missing], b = [missing, 6])
         b = unstack(df, :variable, :value)
-        @test a ≅ b ≅ DataFrame(id = [1, 2], a = [5, missing], b = [missing, 6])
+        @test b ≅ DataFrame(id = [1, 2], id2 = [1, 2], a = [5, missing], b = [missing, 6])
 
         df = DataFrame(id=1:2, variable=["a", "b"], value=3:4)
         @static if VERSION >= v"0.6.0-dev.1980"
@@ -380,6 +434,42 @@ module TestDataFrame
         a = unstack(df, :id, :variable, :value)
         b = unstack(df, :variable, :value)
         @test a ≅ b ≅ DataFrame(id = [1, 2], a = [3, missing], b = [missing, 4])
+
+        df = DataFrame(variable=["x", "x"], value=[missing, missing], id=[1,1])
+        @test_warn "Duplicate entries in unstack at row 2 for key 1 and variable x." unstack(df, :variable, :value)
+        @test_warn "Duplicate entries in unstack at row 2 for key 1 and variable x." unstack(df)
+    end
+
+    @testset "missing values in colkey" begin
+        df = DataFrame(id=[1, 1, 1, missing, missing, missing, 2, 2, 2],
+                       variable=["a", "b", missing, "a", "b", "missing", "a", "b", "missing"],
+                       value=[missing, 2.0, 3.0, 4.0, 5.0, missing, 7.0, missing, 9.0])
+        @test_warn "Missing value in variable variable at row 3. Skipping." unstack(df)
+        udf = unstack(df)
+        @test names(udf) == [:id, :a, :b, :missing]
+        @test udf[:missing] ≅ [missing, 9.0, missing]
+        df = DataFrame(id=[1, 1, 1, missing, missing, missing, 2, 2, 2],
+                       id2=[1, 1, 1, missing, missing, missing, 2, 2, 2],
+                       variable=["a", "b", missing, "a", "b", "missing", "a", "b", "missing"],
+                       value=[missing, 2.0, 3.0, 4.0, 5.0, missing, 7.0, missing, 9.0])
+        @test_warn "Missing value in variable variable at row 3. Skipping." unstack(df, 3, 4)
+        udf = unstack(df, 3, 4)
+        @test names(udf) == [:id, :id2, :a, :b, :missing]
+        @test udf[:missing] ≅ [missing, 9.0, missing]
+    end
+
+    @testset "stack-unstack correctness" begin
+        x = DataFrame(rand(100, 50))
+        x[:id] = [1:99; missing]
+        x[:id2] = string.("a", x[:id])
+        x[:s] = [i % 2 == 0 ? randstring() : missing for i in 1:100]
+        allowmissing!(x, :x1)
+        x[1, :x1] = missing
+        y = melt(x, [:id, :id2])
+        z = unstack(y)
+        @test all(isequal(z[n], x[n]) for n in names(z))
+        z = unstack(y, :variable, :value)
+        @test all(isequal(z[n], x[n]) for n in names(x))
     end
 
     @testset "rename" begin