Changes:
1) allow multiple columns in rowkeys
2) remove incorrect and slower implementation for single column in rowkey