@bkamins 2017-12-05
@nalimilan @cjprybol
Just for a sake of completeness of refactoring `unstack` let me mention that currently we do not handle the following case 100% correct:
```
julia> y = DataFrame(variable=["x","x"], value=[missing,missing], id=[1,1])
2×3 DataFrames.DataFrame
│ Row │ variable │ value   │ id │
├─────┼──────────┼─────────┼────┤
│ 1   │ x        │ missing │ 1  │
│ 2   │ x        │ missing │ 1  │

julia> unstack(y)
1×2 DataFrames.DataFrame
│ Row │ id │ x       │
├─────┼────┼─────────┤
│ 1   │ 1  │ missing │
```
There is a duplicate in `y` that is not reported because it is `missing` (and `missing` is used to indicate yet unfilled data).

The second thing is that `unstack` makes all columns accept missing even if we are not sure that we have to (this means that stack-unstack procedure changes types of columns).

The solution would be to use a temporary data structure that would hold binary indicators if a position is filled (so we have no duplicate `missing` issue). Later those indicators can be also used to decide if we have to make a column accept `Missing` or not.

But I am not sure if you would feel it is worth the effort as it will introduce some memory and computational overhead.