@bkamins 2017-12-05
tests fail because the fix of `allowmissing!` is not tagged yet so we have rerun CI after the update of dependencies.