@nalimilan 2017-12-06
> There is a duplicate in y that is not reported because it is missing (and missing is used to indicate yet unfilled data).
> 
> The second thing is that unstack makes all columns accept missing even if we are not sure that we have to (this means that stack-unstack procedure changes types of columns).
> 
> The solution would be to use a temporary data structure that would hold binary indicators if a position is filled (so we have no duplicate missing issue). Later those indicators can be also used to decide if we have to make a column accept Missing or not.
> 
> But I am not sure if you would feel it is worth the effort as it will introduce some memory and computational overhead.

Why not, I guess the overhead could be kept relatively low. With compiler improvements we should be able to convert an `Array{Union{Missing, T}` to an `Array{T}` very efficiently, and keeping track of whether an entry is missing or not shouldn't be too costly, if you're willing to do the work. Among other things, `ismissing(payload[j][i])` is probably quite slow currently (waiting for more compiler optimizations), and could be much faster if we used a boolean array.

It would make sense to check what other implementations do, though.