@bkamins 2017-12-07
I have fixed correct handling of missing. I have decided to retain that all columns accept missing - this would introduce performance overhead to control for it and I think it is not crucially problematic.