@bkamins 2017-12-07
First part - timing of new vs old code (new code uses `mask` Boolean test):
```
# Old code
julia> @benchmark unstack(y)
  memory estimate:  2.01 MiB
  mean time:        8.176 ms (3.61% GC)
julia> @benchmark unstack(y, :variable, :value)
  memory estimate:  1.85 MiB
  mean time:        6.263 ms (3.67% GC)
julia> @benchmark unstack(z)
  memory estimate:  1.65 MiB
  mean time:        11.022 ms (2.53% GC)
julia> @benchmark unstack(z, :variable, :value)
  memory estimate:  1.92 MiB
  mean time:        10.190 ms (2.86% GC)

# The proposed code with mask array
julia> @benchmark unstack(y)
  memory estimate:  1.85 MiB
  mean time:        6.001 ms (4.54% GC)
julia> @benchmark unstack(y, :variable, :value)
  memory estimate:  1.85 MiB
  mean time:        5.891 ms (3.69% GC)
julia> @benchmark unstack(z)
  memory estimate:  1.92 MiB
  mean time:        9.367 ms (3.08% GC)
julia> @benchmark unstack(z, :variable, :value)
  memory estimate:  1.92 MiB
  mean time:        9.391 ms (2.90% GC)
```