@bkamins 2017-12-07
I performed the profiling and actually it seems that creation of `groupby`:
```
g = groupby(df, setdiff(_names(df), _names(df)[[colkey, value]]), sort=true)
```
is faster than:
```
keycol = CategoricalArray{Union{eltype(df[colkey]), Missing}}(df[colkey])
refkeycol = CategoricalArray{Union{eltype(df[rowkey]), Missing}}(df[rowkey])
```
if they have to be run. This means that `unstack` without `groupby` would be faster if `colkey` and `rowkey` column would be categorial already.

In consequence the question is if we want to make `stack` turn `:variable` column into `Categorical` (it will make `stack` slower but `unstack` faster).