@nalimilan 2017-12-07
> I performed the profiling and actually it seems that creation of groupby:
> 
> g = groupby(df, setdiff(_names(df), _names(df)[[colkey, value]]), sort=true)
> 
> is faster than:
> 
> keycol = CategoricalArray{Union{eltype(df[colkey]), Missing}}(df[colkey])
> refkeycol = CategoricalArray{Union{eltype(df[rowkey]), Missing}}(df[rowkey])
> 
> if they have to be run.

Why not `g = groupby(df, rowkeys, sort=true)` as in the code? Though in this case it should be equivalent?

Also, it's not fair to include `keycol`, since it needs to be created in both cases, right?

> This means that unstack without groupby would be faster if colkey and rowkey column would be categorial already.

But is that indeed the case in benchmarks? In my quick tests, it wasn't, which is really weird.

Anyway, if `groupby` than creating a `CategoricalArray`, that means that hashing a column first, and only then perform dictionary lookups based on these hashes is faster that hashing and doing the lookups on the fly at the same time. Indeed that's the main difference I can think of between the two approaches.