@bkamins 2017-12-07
I was testing on the implementation before my changes (that is the reason for code differences). I have used @profile - those are the lines that came up (and actually single `Categorical` creation was slower).

However, it is weird indeed - I will investigate more into it. What bothers me is:
```
using BenchmarkTools

df = DataFrame(rand(1000,10))
df[:id] = string.("a", 1:1000)
sdf = stack(df)
sdf = sdf[randperm(nrow(sdf)), :]

f1(s) = groupby(s, [:id], sort=true)
f2(s) = CategoricalArray{Union{eltype(s[:id]), Missing}}(s[:id])

@benchmark f1($sdf)
@benchmark f2($sdf)
```

gives:
```
julia> @benchmark f1($sdf)
BenchmarkTools.Trial:
  memory estimate:  601.81 KiB
  allocs estimate:  12377
  --------------
  minimum time:     2.283 ms (0.00% GC)
  median time:      2.422 ms (0.00% GC)
  mean time:        2.528 ms (3.56% GC)
  maximum time:     55.176 ms (94.73% GC)
  --------------
  samples:          1975
  evals/sample:     1

julia> @benchmark f2($sdf)
BenchmarkTools.Trial:
  memory estimate:  756.02 KiB
  allocs estimate:  15795
  --------------
  minimum time:     1.448 ms (0.00% GC)
  median time:      1.549 ms (0.00% GC)
  mean time:        1.693 ms (7.64% GC)
  maximum time:     55.446 ms (97.12% GC)
  --------------
  samples:          2940
  evals/sample:     1
```
which is in line with your intuition and against the results I report above.