@bkamins 2017-12-07
Actually running `@code_warntype` on both implementations is all red. I will check if we can improve something here.