@bkamins 2017-12-07
OK. I have it. We will leave separate implementations for single and multiple columns.
The problem is solved in a standard way by barrier function:
```
function unstack(df::AbstractDataFrame, rowkey::Int, colkey::Int, value::Int)
    refkeycol = CategoricalArray{Union{eltype(df[rowkey]), Missing}}(df[rowkey])
    keycol = CategoricalArray{Union{eltype(df[colkey]), Missing}}(df[colkey])
    _unstack(df, rowkey, colkey, value, refkeycol, keycol)
end
```
(it is a fast writeup - I will push a clean solution when I have it but roughly it is 2x speedup on the test I report above)