@bkamins 2017-12-09
Here are the benchmarks (I have cut the output to leave what is important - timing is given after `unstack` is compiled for given arguments).
The conclusion is that we are 2x faster with this implementation and 0.7 is faster than 0.6 (and keep in mind that the new implementation does more work as it has to handle missing values in `rowkeys` and `colkey`).

Data setup code:
```
x = DataFrame(rand(10000, 1000))
x[:id] = 1:10000
x[:id2] = string.("a", x[:id])
x[:s] = [randstring() for i in 1:10000]
y = melt(x, [:id, :id2])
```

Summary of benchmark results:
```
Julia 0.6

OLD
julia> @time unstack(y);
 10.974269 seconds (121.05 M allocations: 2.254 GiB, 22.68% gc time)
julia> @time unstack(y, :variable, :value);
 10.412161 seconds (73.60 M allocations: 1.859 GiB, 30.93% gc time)

 categorical!(y, [:id, :id2])
julia> @time unstack(y);
 10.393829 seconds (111.06 M allocations: 1.919 GiB, 22.76% gc time)
julia> @time unstack(y, :variable, :value);
 11.879770 seconds (83.64 M allocations: 2.010 GiB, 31.39% gc time)

NEW 
julia> @time unstack(y);
  4.848957 seconds (34.63 M allocations: 1.006 GiB, 42.45% gc time)
julia> @time unstack(y, :variable, :value);
  8.192743 seconds (24.48 M allocations: 1.128 GiB, 56.92% gc time)

 categorical!(y, [:id, :id2])
julia> @time unstack(y);
  4.182978 seconds (24.64 M allocations: 687.415 MiB, 46.12% gc time)
julia> @time unstack(y, :variable, :value);
  9.793470 seconds (34.52 M allocations: 1.280 GiB, 55.00% gc time)

Julia 0.7

OLD
julia> @time unstack(y);
 11.376244 seconds (121.07 M allocations: 2.254 GiB, 11.02% gc time)
julia> @time unstack(y, :variable, :value);
  9.536922 seconds (73.59 M allocations: 1.859 GiB, 22.01% gc time)

 categorical!(y, [:id, :id2])
julia> @time unstack(y);
 10.165981 seconds (111.06 M allocations: 1.919 GiB, 11.03% gc time)
julia> @time unstack(y, :variable, :value);
 10.698634 seconds (83.64 M allocations: 2.010 GiB, 22.48% gc time)

NEW
julia> @time unstack(y);
  4.928271 seconds (34.56 M allocations: 1.005 GiB, 41.12% gc time)
julia>  @time unstack(y, :variable, :value);
  5.809069 seconds (24.48 M allocations: 1.128 GiB, 42.01% gc time)

 categorical!(y, [:id, :id2])
julia> @time unstack(y);
  3.382329 seconds (24.55 M allocations: 686.091 MiB, 28.77% gc time)
julia> @time unstack(y, :variable, :value);
  6.865709 seconds (34.53 M allocations: 1.280 GiB, 40.00% gc time)
```